// src/types/authType.ts
export interface BranchLevels {
  branch_name: string;
}

export interface BranchResponse {
  success: boolean;
  data: BranchDataRes
}

export interface BranchDataRes {
  success:boolean;
  data:BranchLevels[];
}